requirements.txt
flask
gitpython
faiss-cpu
transformers
torch
